--- Assets/Animation/Falling.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7f92543ac8718d54f9069936b648e390
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
