+++ Assets/Animation/PlayerAnimation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b6c570ea42c64646a5919a218349473
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
