+++ Assets/Animation/PlayerAnimation/playerJump.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 845b0a364d42e4e4a997fbbbfa99d2fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
