+++ Assets/Animation/PlayerAnimation/playerMeleeTool.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b55ae9c14528d984bb3c9d17649d5574
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
