+++ Assets/Animation/PlayerAnimation/playerWalk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 75a87b9844f68e8428fbb48655516c68
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
