+++ Assets/Animation/PlayerAnimation/playeridle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e40729c07d592543af330003ca616df
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
