--- Assets/Animation/playerControl.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e9cfc99461169ac4d93dc54053f6c062
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
