--- Assets/Animation/playerRun.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2ce6bd171e94d8640a3765f8811d9080
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
