--- Assets/Animation/playerWalk.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 75a87b9844f68e8428fbb48655516c68
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
