--- Assets/PlayerInteraction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0164f656bb0f9584e90898f2cbbda03a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
