--- Assets/playerFall.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 72472cc4509ad1f4488244f07dfab5af
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
